src/app/api/ai/embed/audio/requirements.txt
pyannote.audio==3.3.1
pydub==0.25.1
openai==1.37.0
python-dotenv==1.0.1
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
numpy==1.26.4